backend/requirements.txt
flask==3.0.0
firebase-admin==6.2.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0